fix(core): Clear inherited config for called callbacks (#7174)
