+++ langchain-core/src/callbacks/tests/callbacks.test.ts
@@ -1,6 +1,7 @@
 /* eslint-disable no-promise-executor-return */
 import { test, expect } from "@jest/globals";
 import * as uuid from "uuid";
+import { AsyncLocalStorage } from "node:async_hooks";
 import { CallbackManager } from "../manager.js";
 import { BaseCallbackHandler, type BaseCallbackHandlerInput } from "../base.js";
 import type { Serialized } from "../../load/serializable.js";
@@ -10,6 +11,8 @@ import type { AgentAction, AgentFinish } from "../../agents.js";
 import { BaseMessage, HumanMessage } from "../../messages/index.js";
 import type { LLMResult } from "../../outputs.js";
 import { RunnableLambda } from "../../runnables/base.js";
+import { AsyncLocalStorageProviderSingleton } from "../../singletons/index.js";
+import { awaitAllCallbacks } from "../promises.js";
 
 class FakeCallbackHandler extends BaseCallbackHandler {
   name = `fake-${uuid.v4()}`;
@@ -535,4 +538,39 @@ test("chain should still run if a normal callback handler throws an error", asyn
     }
   );
   expect(res).toEqual("hello world");
+});
+
+test("runnables in callbacks should be root runs", async () => {
+  AsyncLocalStorageProviderSingleton.initializeGlobalInstance(
+    new AsyncLocalStorage()
+  );
+  const nestedChain = RunnableLambda.from(async () => {
+    const subRun = RunnableLambda.from(async () => "hello world");
+    return await subRun.invoke({ foo: "bar" });
+  });
+  let error;
+  let finalInputs;
+  const res = await nestedChain.invoke(
+    {},
+    {
+      callbacks: [
+        {
+          handleChainStart: (_chain, inputs) => {
+            finalInputs = inputs;
+            try {
+              expect(
+                AsyncLocalStorageProviderSingleton.getRunnableConfig()
+              ).toEqual(undefined);
+            } catch (e) {
+              error = e;
+            }
+          },
+        },
+      ],
+    }
+  );
+  await awaitAllCallbacks();
+  expect(res).toEqual("hello world");
+  expect(error).toBe(undefined);
+  expect(finalInputs).toEqual({ foo: "bar" });
 });
